massive restructure to the python package
