+++ .gitignore
@@ -7,3 +7,4 @@ my_data
 __pycache__
 *.pyc
 venv
+.vscode
