--- snowflake_data_profiler/.flaskenv
@@ -1,1 +1,0 @@
-FLASK_APP=profiler-app.py
