+++ snowflake_data_profiler/config.py
@@ -1,5 +1,4 @@
 import os
 
-
 class Config(object):
     SECRET_KEY = os.environ.get('SECRET_KEY') or 'you-will-never-guess'
