--- snowflake_data_profiler/profiler-app.py
@@ -1,5 +1,0 @@
-from flask import Flask
-from ..config import Config
-
-app = Flask(__name__)
-app.config.from_object(Config)
