+++ snowflake_data_profiler/requirements.txt
@@ -3,7 +3,6 @@ pandas-profiling==2.8.0
 Flask==1.1.2
 Flask-WTF==0.14.3
 Jinja2==2.11.2
-pysnowflake==0.1.3
 snowflake-connector-python==2.2.8
 WTForms==2.3.1
